requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Inject==4.3.1
pyaes==1.6.1
pyasn1==0.4.8
pymitter==0.3.0
pyTelegramBotAPI==3.7.6
python-jsonstore==1.3.0
requests==2.25.1
rsa==4.7.2
Telethon==1.21.1
typeguard==2.12.0
urllib3==1.26.4
tengi
